# EventManager/app/docs/swagger.yaml
basePath: /api/event-manager
definitions:
  hateoas.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
      title:
        type: string
    type: object
...
  httpdto.HttpCreateEventPacketInclusion:
    type: object
...
  httpdto.HttpResponseEventPacketInclusion:
    properties:
      _links:
        additionalProperties:
          $ref: '#/definitions/hateoas.Link'
        type: object
      id_event:
        type: integer
      id_packet:
        type: integer
    type: object
...
  httpdto.HttpUpdateEventPacketInclusion:
    type: object
...
host: localhost:12345
info:
  contact: {}
  description: Event management service for the POS system. Handles events, event
    packets, tickets, and inclusions.
  title: EventManager Service API
  version: "1.0"
paths:
  /event-packet-inclusions/event/{event_id}:
    get:
      consumes:
      - application/json
      description: Retrieve all event packets linked to a specific event
      parameters:
      - description: Bearer token (optional)
        in: header
        name: Authorization
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of event packets
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid event ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all packet inclusions for an event
      tags:
      - event-packet-inclusions
  /event-packet-inclusions/event/{event_id}/packet/{packet_id}:
    delete:
      consumes:
      - application/json
      description: Remove the link between an event and an event packet
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: integer
      - description: Packet ID
        in: path
        name: packet_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Inclusion deleted successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEventPacketInclusion'
        "400":
          description: Invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Inclusion not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete event packet inclusion
      tags:
      - event-packet-inclusions
    patch:
      consumes:
      - application/json
      description: Update the details of an event-packet inclusion relationship
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: integer
      - description: Packet ID
        in: path
        name: packet_id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: inclusion
        required: true
        schema:
          $ref: '#/definitions/httpdto.HttpUpdateEventPacketInclusion'
      produces:
      - application/json
      responses:
        "200":
          description: Inclusion updated successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEventPacketInclusion'
        "400":
          description: Invalid parameters or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Inclusion not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update event packet inclusion
      tags:
      - event-packet-inclusions
    post:
      consumes:
      - application/json
      description: Link an event packet to an event (many-to-many relationship)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: integer
      - description: Packet ID
        in: path
        name: packet_id
        required: true
        type: integer
      - description: Inclusion details
        in: body
        name: inclusion
        required: true
        schema:
          $ref: '#/definitions/httpdto.HttpCreateEventPacketInclusion'
      produces:
      - application/json
      responses:
        "201":
          description: Inclusion created successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEventPacketInclusion'
        "400":
          description: Invalid parameters or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event or packet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Inclusion already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create event packet inclusion
      tags:
      - event-packet-inclusions
  /event-packet-inclusions/packet/{packet_id}:
    get:
      consumes:
      - application/json
      description: Retrieve all events linked to a specific event packet
      parameters:
      - description: Bearer token (optional)
        in: header
        name: Authorization
        type: string
      - description: Packet ID
        in: path
        name: packet_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of events
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid packet ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Packet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all event inclusions for a packet
      tags:
      - event-packet-inclusions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"